handle invalid queries
