+++ src/components/Search/SearchPageHeader/SearchPageHeader.tsx
@@ -1,18 +1,13 @@
-import React, {useCallback, useMemo} from 'react';
+import React from 'react';
 import {View} from 'react-native';
 import {useOnyx} from 'react-native-onyx';
 import type {DropdownOption} from '@components/ButtonWithDropdownMenu/types';
 import {useSearchContext} from '@components/Search/SearchContext';
 import type {SearchQueryJSON} from '@components/Search/types';
-import TextInputClearButton from '@components/TextInput/TextInputClearButton';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
-import {clearAdvancedFilters} from '@libs/actions/Search';
-import Navigation from '@libs/Navigation/Navigation';
-import {buildCannedSearchQuery} from '@libs/SearchQueryUtils';
 import SearchSelectedNarrow from '@pages/Search/SearchSelectedNarrow';
 import type CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
-import ROUTES from '@src/ROUTES';
 import type DeepValueOf from '@src/types/utils/DeepValueOf';
 import SearchPageHeaderInput from './SearchPageHeaderInput';
 
@@ -34,19 +29,6 @@ function SearchPageHeader({queryJSON, searchRouterListVisible, hideSearchRouterL
 
     const selectedTransactionsKeys = Object.keys(selectedTransactions ?? {});
 
-    const clearFilters = useCallback(() => {
-        clearAdvancedFilters();
-        Navigation.navigate(
-            ROUTES.SEARCH_ROOT.getRoute({
-                query: buildCannedSearchQuery(),
-            }),
-        );
-    }, []);
-
-    const InputRightComponent = useMemo(() => {
-        return <TextInputClearButton onPressButton={clearFilters} />;
-    }, [clearFilters]);
-
     if (shouldUseNarrowLayout && selectionMode?.isEnabled) {
         return (
             <View>
@@ -64,7 +46,6 @@ function SearchPageHeader({queryJSON, searchRouterListVisible, hideSearchRouterL
             onSearchRouterFocus={onSearchRouterFocus}
             queryJSON={queryJSON}
             hideSearchRouterList={hideSearchRouterList}
-            inputRightComponent={queryJSON.inputQuery === buildCannedSearchQuery() ? undefined : InputRightComponent}
             handleSearch={handleSearch}
         />
     );
